+++ history.md
@@ -7,6 +7,7 @@
 	* bugfix: history date cut off
     * Toast on change value to find stray clicks
     * new items private by default
+    * Edit description should shows context: date, name, value
     * bugfix: Item menu displayed at top is small.
     * bugfix: Category selector too tall
     * bugfix: applies to time was utc - now displays and edits in local time